+++ src/com/ansorgit/plugins/bash/runner/BashConfigurationType.java
@@ -66,9 +66,8 @@ public class BashConfigurationType extends ConfigurationTypeBase implements Poss
         @NotNull
         @Override
         public RunConfiguration createTemplateConfiguration(@NotNull Project project) {
-            BashRunConfiguration configuration = new BashRunConfiguration(new RunConfigurationModule(project), this, "");
-            configuration.setInterpreterPath(BashInterpreterDetection.instance().findBestLocation());
-
+            BashRunConfiguration configuration = new BashRunConfiguration("", new RunConfigurationModule(project), this);
+            configuration.setUseProjectInterpreter(true);
             return configuration;
         }
     }