+++ src/com/ansorgit/plugins/bash/runner/BashRunConfigProducer.java
@@ -24,11 +24,11 @@ import com.intellij.execution.actions.RunConfigurationProducer;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.util.Ref;
 import com.intellij.openapi.util.io.FileUtil;
-import com.intellij.openapi.util.text.StringUtil;
+import com.intellij.openapi.util.io.FileUtilRt;
+import com.intellij.openapi.vfs.VfsUtilCore;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
-import org.apache.commons.lang.StringUtils;
 
 /**
  * Bash run config producer which looks at the current context to create a new run configuation.
@@ -51,22 +51,21 @@ public class BashRunConfigProducer extends RunConfigurationProducer<BashRunConfi
         }
 
         PsiFile psiFile = psiElement.getContainingFile();
-        if (psiFile == null || !(psiFile instanceof BashFile)) {
+        if (!(psiFile instanceof BashFile)) {
             return false;
         }
+        sourceElement.set(psiFile);
 
         VirtualFile file = location.getVirtualFile();
         if (file == null) {
             return false;
         }
 
-        sourceElement.set(psiFile);
-
-        configuration.setName(location.getVirtualFile().getPresentableName());
-        configuration.setScriptName(file.getPath());
+        configuration.setName(file.getPresentableName());
+        configuration.setScriptName(VfsUtilCore.virtualToIoFile(file).getAbsolutePath());
 
         if (file.getParent() != null) {
-            configuration.setWorkingDirectory(file.getParent().getPath());
+            configuration.setWorkingDirectory(VfsUtilCore.virtualToIoFile(file.getParent()).getAbsolutePath());
         }
 
         Module module = context.getModule();
@@ -86,12 +85,6 @@ public class BashRunConfigProducer extends RunConfigurationProducer<BashRunConfi
 
                 configuration.setInterpreterOptions(shebang.shellCommandParams());
             }
-        }
-
-        //fallback location if none was found
-        if (StringUtil.isEmptyOrSpaces(configuration.getInterpreterPath())) {
-            String bashPath = BashInterpreterDetection.instance().findBestLocation();
-            configuration.setInterpreterPath(StringUtils.trimToEmpty(bashPath));
         }
 
         return true;