+++ src/com/ansorgit/plugins/bash/runner/BashRunConfigurationEditor.java
@@ -33,13 +33,13 @@ class BashRunConfigurationEditor extends SettingsEditor<BashRunConfiguration> {
     @Override
     protected void resetEditorFrom(BashRunConfiguration runConfiguration) {
         form.reset(runConfiguration);
-        form.resetBash(runConfiguration);
+        form.resetFormTo(runConfiguration);
     }
 
     @Override
     protected void applyEditorTo(BashRunConfiguration runConfiguration) throws ConfigurationException {
         form.applyTo(runConfiguration);
-        form.applyBashTo(runConfiguration);
+        form.applySettingsTo(runConfiguration);
     }
 
     @Override